VMS-8355: When user changes secondary stream resolution for Hikvision, wrong resolution is set on camera.
<==== Mercurial data ====>
Branch: vms_3.2_dev

+++ mediaserver_core/src/nx/mediaserver/resource/camera_advanced_parameters_providers.cpp
@@ -153,7 +153,7 @@ void StreamCapabilityAdvancedParametersProvider::updateMediaCapabilities()
     auto streamCapabilities = m_capabilities.value(key);
 
     auto mediaCapabilities = m_camera->cameraMediaCapability();
-    mediaCapabilities.streamCapabilities[(int)m_streamIndex] = streamCapabilities;
+    mediaCapabilities.streamCapabilities[m_streamIndex] = streamCapabilities;
     m_camera->setCameraMediaCapability(mediaCapabilities);
 }
